Create function for setting completed status on games
Add event listener to completed button that calls setCompletedStatus on click. The setCompletedStatus function checks for the current status of completed insidethe game objects properties. Depending on the current status it is changed from true to false or the opposite case.

Changed the way the program checks to see if the user completed the game or not.When user types in true the string is compared to the boolean and re-assigned the completed variable. This fixes the setCompletedStatus comparing a string to a boolean.

Added semicolons to lines.