Change the way games are added to the library
Remove the push method and replace with unshift so that the new game is always at position 0. After a new game is created we then add that new game to the display with the displayLibrary function.

Remove the for loop in the displayLibrary function because we no longer create new books all at once. We use the form which means new games are added one by one. Using this we just change the index we are displaying to 0 because the game that needs to be added is always at position 0.